requirements.txt
Flask
google-generativeai
flask-cors
pandas
scikit-learn
joblib
numpy